added clearable functionablity. updated styles
