+++ TimePickerComponent/.eslintrc.json
@@ -3,7 +3,7 @@
     "browser": true,
     "es2021": true
   },
-  "extends": ["eslint:recommended", "plugin:react/recommended"],
+  "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended", "plugin:react/recommended", "prettier"],
   "globals": {
     "ComponentFramework": true
   },
@@ -14,11 +14,22 @@
   },
   "plugins": ["@microsoft/power-apps", "@typescript-eslint"],
   "rules": {
-    "no-unused-vars": "off"
+    "no-unused-vars": "off",
+    "no-explicit-any": "off",
+    "@typescript-eslint/no-unused-vars": "off",
+    "@typescript-eslint/no-explicit-any": "off"
   },
   "settings": {
     "react": {
       "version": "detect"
     }
-  }
+  },
+  "overrides": [
+    {
+      "files": ["*.ts"],
+      "rules": {
+        "camelcase": [2, { "properties": "never" }]
+      }
+    }
+  ]
 }
